+++ virtualFittingApp/src/pages/my/index.ts
@@ -6,3 +6,4 @@ export * from "@/pages/my/ui/MyCancel";
 export * from "@/pages/my/ui/MyLike";
 export * from "@/pages/my/utils/imagePreview";
 export * from "@/pages/my/utils/imageUpload";
+export * from "@/pages/my/ui/OrderContentList";
